bcm53xx: image: sync targets names with DT compatible
Some devices profile names and *.dts files are named without suffix
"-v1" but the DT compatible has the suffix. (xwr-3150 exception)

This breaks ASU sysupgrades since the default SUPPORTED_DEVICES is the
makefile image profile target and it does not match the actual DT
compatible sent by the ASU client.

This is not breaking sysupgrade because the platform implementation
does not depends on DEVICE_NAME(not use sysupgrade-TAR). And the
board_name() checks are done against the correct DT compatible.[1]

[1]: https://github.com/openwrt/openwrt/blob/09e7b24ceb5d9310334c4f88327dc6d58052e7fd/target/linux/bcm53xx/base-files/lib/upgrade/platform.sh#L36-L62
Fixes: https://github.com/openwrt/asu/issues/560
Fixes: https://github.com/openwrt/asu/issues/560#issuecomment-2083236275
Fixes: https://github.com/openwrt/asu/issues/1108
Fixes: https://forum.openwrt.org/t/luci-attended-sysupgrade-support-thread/230552/243
Fixes: c459a6b "bcm53xx: add support for Luxul FullMAC WiFi devices"
Fixes: 1d47f81 "bcm53xx: build images for Luxul ABR-4500 and XBR-4500 routers"
Signed-off-by: Mario Andrés Pérez <mapb_@outlook.com>
Link: https://github.com/openwrt/openwrt/pull/21056
Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>